Full Engine Source

+++ Engine/msvc10/Camera.cpp
@@ -1,12 +1,15 @@
 #include "Advanced2D.h"
 
+
+//***mod--floats to doubles
+
 Camera::Camera(void)
 {
 	p_position = D3DXVECTOR3(0.0f,0.0f,10.0f);
 	p_updir = D3DXVECTOR3(0.0f,1.0f,0.0f);
 
 	//hard coded to 1.3333 by default
-	float ratio = 640 / 480;
+	double ratio = 640 / 480;
 	setPerspective(3.14159f / 4, ratio, 1.0f, 2000.0f);
 
 }
@@ -15,7 +18,7 @@ Camera::~Camera(void)
 {
 }
 
-void Camera::setPerspective(float fov, float aspectRatio, float nearRange, float farRange)
+void Camera::setPerspective(double fov, double aspectRatio, double nearRange, double farRange)
 {
 	this->setFOV(fov);
 	this->setAspectRatio(aspectRatio);
@@ -27,7 +30,8 @@ void Camera::setPerspective(float fov, float aspectRatio, float nearRange, float
 void Camera::Update()
 {
     //set the camera's perspective matrix
-    D3DXMatrixPerspectiveFovLH(&this->p_matrixProj, this->p_fov, this->p_aspectRatio, this->p_nearRange, this->p_farRange);
+    D3DXMatrixPerspectiveFovLH(&this->p_matrixProj, (float)this->p_fov, 
+		(float)this->p_aspectRatio, (float)this->p_nearRange, (float)this->p_farRange);
     g_engine->getDevice()->SetTransform(D3DTS_PROJECTION, &this->p_matrixProj);
 
     //set the camera's view matrix
@@ -35,11 +39,11 @@ void Camera::Update()
     g_engine->getDevice()->SetTransform(D3DTS_VIEW, &this->p_matrixView);
 }
 
-void Camera::setPosition(float x, float y, float z)
+void Camera::setPosition(double x, double y, double z)
 {
-	this->p_position.x = x;
-	this->p_position.y = y;
-	this->p_position.z = z;
+	this->p_position.x = (float)x;
+	this->p_position.y = (float)y;
+	this->p_position.z = (float)z;
 }
 
 void Camera::setPosition(D3DXVECTOR3 position)
